ref: modularizing solvers
