+++ src/lib/machine/reserved.ts
@@ -2,3 +2,5 @@ export const SAT_STATE_ID = 999;
 export const UNSAT_STATE_ID = 666;
 export const DECIDE_STATE_ID = 333;
 export const BACKTRACKING_STATE_ID = 444;
+
+
